requirements.txt
requests~=2.32.3
allure-python-commons~=2.13.5
pytest~=8.3.4